refactor : profile액티비티 Register액티비티와 병합 후 삭제
