+++ app/src/main/AndroidManifest.xml
@@ -16,13 +16,6 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".ui.screen.login.ProfileActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".ui.screen.SplashActivity"
             android:exported="true">
             <intent-filter>
