+++ app/src/main/java/com/example/woomansi/ui/screen/SplashActivity.java
@@ -9,9 +9,7 @@ import android.os.Handler;
 import com.example.woomansi.R;
 import com.example.woomansi.data.model.UserModel;
 import com.example.woomansi.ui.screen.login.LoginActivity;
-import com.example.woomansi.ui.screen.login.ProfileActivity;
 import com.example.woomansi.ui.screen.main.MainActivity;
-import com.example.woomansi.util.SharedPreferencesUtil;
 import com.example.woomansi.util.UserCache;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
